Give every target the VC6 shim, not just the DLL
`tests/recovery_fixtures.h:35` is `#include <cstdint>`, and building
`recovery-gameplay-tests` under the vc6 lane died on
`C1083: Cannot open include file: 'cstdint'`.

The include is correct and stays. The tree already solved this once, as a shim
directory rather than an edit to each includer - `cmake/vc6-include/cstdint`
and `cmake/vc6-include/stdint.h`, each ending in `#include "vc6_compat.h"`,
with `stdint.h`'s own comment stating the rule: the source keeps saying the
standard thing and there is nothing to undo when this compiler is retired.

What was wrong is that the shim was attached to `OpenSMACX` with
`target_include_directories`, i.e. a hand-maintained list of one. Every target
added afterwards that spelled `<cstdint>` broke identically, and one did. It is
now `include_directories` before the first target, so the whole directory
inherits it.

The same defect had a second half. The C++11 spellings `vc6_compat.h` supplies
- `nullptr`, `constexpr`, `static_assert`, the `namespace std`
using-declarations - reach the DLL through `src/stdafx.h`, which every DLL
source includes. The test targets have no such header, so `globals-diff-tests`
and `lifted-oracle-fold-tests` failed on `'std' : is not a class or namespace
name` and `'nullptr' : undeclared identifier`. A directory-scope
`/FI src/vc6_compat.h` delivers it everywhere; `#pragma once` makes the double
delivery to stdafx.h sources a no-op and `tools/vc6_cl.py` already translates
`/FI`.

Then the source, all of it valid C++98 AND C++11 so the MinGW side is
unaffected:
  * `tests/recovery_fixtures.h` - `long long` to `int64_t`. cl 12.00.8168
    rejects the former with `C2632: 'long' followed by 'long' is illegal`, and
    those two lines were the only `long long` in the tree, which is proof this
    header had never been through that compiler.
  * `tests/recovery_gameplay_tests.cpp` - an alias declaration to a typedef,
    `>>` to `> >`, `emplace_back` to `push_back(make_pair(...))`, and a
    range-for to an iterator loop. Those four cost 102 errors; only three were
    real and the other 99 were the parser desynchronising, including 42
    `C2078: too many initializers` aimed at `int ids[] = {0, 1};` aggregates
    that were correct all along.
  * `src/vc6_compat.h` - `srand` and `rand` added to the `namespace std`
    block, which `byte_match_census.std_shim()` also reads.

`OpenSMACX.dll` and `recovery-gameplay-tests.exe` now build in
`build/vc6-release`; neither test executable had ever been built there.

NOT DONE, and it is a project rather than a fix: `recovery-leaf-tests` needs
`<array>`, and the leaf harness is C++11 throughout. The runtime-verification
suite has never been compiled by this compiler, so `ctest` in the vc6 lanes
cannot be green yet.

decomp_status --check: 1267 claims, 0 not reproduced.

+++ tests/recovery_fixtures.h
@@ -56,8 +56,14 @@ inline int &failure_count() {
  * thrown away, so no commit contains it and `git log -S` finds nothing. A
  * proof nobody can re-run is a claim.
  */
-inline long long &expect_evaluations() {
-    static long long evaluated = 0;
+// `int64_t`, not `long long`: cl 12.00.8168 rejects the latter outright with
+// `C2632: 'long' followed by 'long' is illegal` (src/vc6_compat.h:69-70 records
+// the same measurement), and these two lines were the only `long long` in the
+// tree - proof this header had never been through that compiler. `int64_t` is
+// 64-bit everywhere, and vc6_compat.h typedefs it to `__int64` for VC6, so the
+// counter keeps its width on every toolchain and needs no `#if`.
+inline int64_t &expect_evaluations() {
+    static int64_t evaluated = 0;
     return evaluated;
 }
